+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Symbol names in stack trace
 - Uses source references for disassembled sources, rather than writing to temp files
+- Improved memory expressions
 
 ## [0.6.0] - 2022-06-06